Change the directory structure for easier maintenance
