--- sqlite-wasm-rs/src/fragile.rs
@@ -1,36 +1,0 @@
-//! A [`FragileComfirmed<T>`] wraps a non sendable `T` to be safely send to other threads.
-//!
-//! Once the value has been wrapped it can be sent to other threads but access
-//! to the value on those threads will fail.
-
-use std::ops::{Deref, DerefMut};
-
-use fragile::Fragile;
-
-pub struct FragileComfirmed<T> {
-    fragile: Fragile<T>,
-}
-
-unsafe impl<T> Send for FragileComfirmed<T> {}
-unsafe impl<T> Sync for FragileComfirmed<T> {}
-
-impl<T> FragileComfirmed<T> {
-    pub fn new(t: T) -> Self {
-        FragileComfirmed {
-            fragile: Fragile::new(t),
-        }
-    }
-}
-
-impl<T> Deref for FragileComfirmed<T> {
-    type Target = T;
-    fn deref(&self) -> &Self::Target {
-        self.fragile.get()
-    }
-}
-
-impl<T> DerefMut for FragileComfirmed<T> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.fragile.get_mut()
-    }
-}
