+++ sqlite-wasm-rs/src/impl.rs
