+++ sqlite-wasm-rs/src/libsqlite3/bindings.rs
