+++ sqlite-wasm-rs/src/libsqlite3/error.rs
